pyproject.toml
[tool.poetry]
name = "aaactl"
repository = "https://github.com/fullctl/aaactl"
version = "0.1.0"
description = "account control"
authors = ["20C <code@20c.com>"]
license = "Apache-2.0"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
]


[tool.poetry.scripts]
aaactl = "aaactl.cli:main"


[tool.poetry.dependencies]
python = "^3.9"
django-reversion = "<3"
djangorestframework = "<4,>=3"
Django = "<3"
confu = ">=1.5,<2"
social-auth-app-django = ">3,<4"
psycopg2-binary = ">2.8,<3"
django-crispy-forms = "<2,>=1.8"
django-recaptcha = "<3,>=2"
django-handleref = "<1,>=0.5"
stripe = "*"
python-dateutil = "*"
django-countries = "*"
django-oauth-toolkit = "*"
django-cors-middleware = "*"
pytest-filedata = "<1"
django-grainy = ">1.7,<2" 
grainy = "<2"
fullctl = { url = "https://github.com/fullctl/fullctl/archive/master.zip" }

[tool.poetry.dev-dependencies]
# tests
pytest = ">=6.0.1"
pytest-django = ">=3.8.0"
pytest-cov = ">=2.7.1, <3"
pytest-mock = ">=3.2.0"
codecov = ">=2.0.15, <3"
coverage = ">=5, <6"

# linting
bandit = "^1.6.2"
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
